MQTT-BDT-main/requirements.txt
flask
numpy
paho-mqtt
pyyaml
requests
pymongo